requirements.txt
vobject>=0.9
caldav>=0.8
python-dotenv>=0.17
icalendar>=4.0
schedule>=1.1
recurring-ical-events
python-dateutil>=2.8
pandas>=1.2
sqlalchemy>=1.4
pymysql>=1.0
requests>=2.25
flask>=1.1
cachetools>=4.2
werkzeug>=1.0
gunicorn>=20.1
numpy>=1.20
pytest>=6.2
pytz
intervaltree
